+++ packages/dart/lib/src/protocol/sentry_geo.dart
@@ -1,12 +1,20 @@
 /// Geographical location of the end user or device.
 class SentryGeo {
-  SentryGeo({this.city, this.countryCode, this.region});
+  SentryGeo({
+    this.city,
+    this.countryCode,
+    this.region,
+    this.subregion,
+    this.subdivision,
+  });
 
   factory SentryGeo.fromJson(Map<String, dynamic> json) {
     return SentryGeo(
       city: json['city'],
       countryCode: json['country_code'],
       region: json['region'],
+      subregion: json['subregion'],
+      subdivision: json['subdivision'],
     );
   }
 
@@ -19,11 +27,19 @@ class SentryGeo {
   /// Human readable region name or code.
   final String? region;
 
+  /// Subregion (e.g. a continental area).
+  final String? subregion;
+
+  /// Subdivision (e.g. state, province).
+  final String? subdivision;
+
   Map<String, dynamic> toJson() {
     return {
       if (city != null) 'city': city,
       if (countryCode != null) 'country_code': countryCode,
       if (region != null) 'region': region,
+      if (subregion != null) 'subregion': subregion,
+      if (subdivision != null) 'subdivision': subdivision,
     };
   }
 }