+++ packages/dart/lib/src/sentry_envelope_item.dart
@@ -94,6 +94,21 @@ class SentryEnvelopeItem {
     );
   }
 
+  /// Create a [SentryEnvelopeItem] which holds pre-encoded span data.
+  /// This is used by the spans buffer to send pre-encoded spans.
+  @internal
+  factory SentryEnvelopeItem.fromSpansData(List<int> payload, int spansCount) {
+    return SentryEnvelopeItem(
+      SentryEnvelopeItemHeader(
+        SentryItemType.span,
+        itemCount: spansCount,
+        contentType: 'application/vnd.sentry.items.span.v2+json',
+      ),
+      () => payload,
+      originalObject: null,
+    );
+  }
+
   /// Create a [SentryEnvelopeItem] which holds pre-encoded metric data.
   /// This is used by the buffer to send pre-encoded metric batches.
   @internal