+++ packages/flutter/lib/src/feedback/sentry_feedback_options.dart
@@ -1,3 +1,6 @@
+import 'dart:ui';
+import '../../sentry_flutter.dart';
+
 class SentryFeedbackOptions {
   // Form Configuration
 
@@ -45,6 +48,9 @@ class SentryFeedbackOptions {
   /// The message displayed after a successful feedback submission.
   String successMessageText = 'Thank you for your report!';
 
+  /// Whether to show the default success message after a successful feedback submission.
+  bool showSuccessMessage = true;
+
   /// The label next to the name input field.
   String nameLabel = 'Name';
 
@@ -74,4 +80,14 @@ class SentryFeedbackOptions {
 
   /// The label of the take screenshot button shown outside of the feedback widget.
   String takeScreenshotButtonLabel = 'Take Screenshot';
+
+  // Theme Configuration
+
+  /// Background color used for the success snack bar.
+  Color successColor = const Color(0xFF268D75);
+
+  // Event Callbacks
+
+  /// Called when feedback is successfully submitted.
+  void Function(SentryFeedback feedback, SentryId eventId)? onSubmitSuccess;
 }